requirements.txt
python-dotenv
langchain_community
langchain_core
langchain_openai
openai